+++ docs/release-notes.md
@@ -12,6 +12,8 @@ Major changes:
 
 Minor changes:
 
+- Hetzner: fix duplicate attribute prefix
+
 Packaging changes: